refactor(ui): implement dynamic imports and split navbar components - add dynamic imports with loading states for landing page components (LogoCloud, GlobalUptime, CodeExample, Features, Cta) - replace 3D globe with CSS radial gradient in GlobalUptime for performance - extract Navbar logic into DesktopNav and MobileNav components - add TRPCProvider layouts for auth, dashboard, and status routes - simplify root layout and dashboard token handling
