+++ apps/client/app/(auth)/layout.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { TRPCProvider } from "../providers";
+
+export default function AuthLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <TRPCProvider>{children}</TRPCProvider>;
+}
