+++ apps/client/app/components/ui/GlobalDatabase.tsx
@@ -1,60 +1,22 @@
-"use client";
-import createGlobe from "cobe";
-import { FunctionComponent, useEffect, useRef } from "react";
+import type { FunctionComponent } from "react";
+
+const features = [
+  {
+    name: "Global Clusters",
+    description: "Enable low-latency global access, enhancing performance.",
+  },
+  {
+    name: "Serverless Triggers",
+    description: "Trigger functions automatically for dynamic app behavior.",
+  },
+  {
+    name: "Monitoring & Alerts",
+    description:
+      "Monitor health with key metrics or integrate third-party tools.",
+  },
+];
 
 export const GlobalUptime: FunctionComponent = () => {
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-
-  useEffect(() => {
-    let phi = 4.7;
-
-    const globe = createGlobe(canvasRef.current!, {
-      devicePixelRatio: 2,
-      width: 1200 * 2,
-      height: 1200 * 2,
-      phi: 0,
-      theta: -0.3,
-      dark: 1,
-      diffuse: 1.2,
-      mapSamples: 25000,
-      mapBrightness: 13,
-      mapBaseBrightness: 0.05,
-      baseColor: [0.3, 0.3, 0.3],
-      glowColor: [0.15, 0.15, 0.15],
-      markerColor: [100, 100, 100],
-      markers: [
-        // { location: [37.7595, -122.4367], size: 0.03 }, // San Francisco
-        // { location: [40.7128, -74.006], size: 0.03 }, // New York City
-        // { location: [35.6895, 139.6917], size: 0.03 }, // Tokyo
-        // { location: [28.7041, 77.1025], size: 0.03 }, // Delhi
-      ],
-      onRender: (state: { phi?: number }) => {
-        state.phi = phi;
-        phi += 0.0002;
-      },
-    });
-
-    return () => {
-      globe.destroy();
-    };
-  }, []);
-
-  const features = [
-    {
-      name: "Global Clusters",
-      description: "Enable low-latency global access, enhancing performance.",
-    },
-    {
-      name: "Serverless Triggers",
-      description: "Trigger functions automatically for dynamic app behavior.",
-    },
-    {
-      name: "Monitoring & Alerts",
-      description:
-        "Monitor health with key metrics or integrate third-party tools.",
-    },
-  ];
-
   return (
     <div className="px-3">
       <section
@@ -73,11 +35,9 @@ export const GlobalUptime: FunctionComponent = () => {
         >
           The Global <br /> Uptime Monitor
         </h2>
-        <canvas
-          className="absolute top-[7.1rem] z-20 aspect-square size-full max-w-fit md:top-[12rem]"
-          ref={canvasRef}
-          style={{ width: 1200, height: 1200 }}
-        />
+        <div className="absolute top-[7.1rem] z-20 aspect-square size-full max-w-fit md:top-[12rem]">
+          <div className="h-full w-full rounded-full bg-[radial-gradient(circle_at_30%_20%,theme(colors.indigo.300),transparent_60%),radial-gradient(circle_at_70%_80%,theme(colors.indigo.700),transparent_55%)] opacity-80" />
+        </div>
         <div className="z-20 -mt-32 h-[36rem] w-full overflow-hidden md:-mt-36">
           <div className="absolute bottom-0 h-3/5 w-full bg-gradient-to-b from-transparent via-gray-950/95 to-gray-950" />
           <div className="absolute inset-x-6 bottom-12 m-auto max-w-4xl md:top-2/3">
