+++ apps/client/app/dashboard/layout.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { TRPCProvider } from "../providers";
+
+export default function DashboardLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <TRPCProvider>{children}</TRPCProvider>;
+}
