+++ apps/client/app/status/layout.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { TRPCProvider } from "../providers";
+
+export default function StatusLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <TRPCProvider>{children}</TRPCProvider>;
+}
